Sort bags by numeric version, not lexicographic version
One of the advantages of building our own browser is that we recognise
the first-class types of the storage service, one of which is numeric
versions.  In the S3 console, these sort lexicographically,
e.g. v1, v10, v11, v2, v3, …

By adding a second pass of sorting before we render the results, we can
display versions in the correct order.