Remove an unused route/method
This route was present when I thought getting the file tallies would be
more expensive, and I thought there might be a "get file types" button.
This route was meant to return the data about file extensions, e.g.

    {".xml": 5, ".jpg": 3}

But now that data is bundled in the main payload, so this route isn't
needed (and isn't used anywhere).